Destroy the rest of the mizu types
